requirements.txt
APScheduler==3.10.4
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
TgCrypto==1.2.5
tzdata==2024.1
tzlocal==5.2
